Header.vue の表示オプション選択UIを OptionSelector.vue に切り出した
